+++ CHANGELOG.md
@@ -13,3 +13,8 @@
 
 - Add similarly named buffers filters
 - Always show current project filter and error on selection if `projectile` is not installed
+
+\*\*\* 0.1.4
+
+- Add `frog-jump-buffer-default-filters-capital-letters` as an option instead of numbers.
+- Allow a simpler method of adding custom filter actions and opting out of the default ones.